Fixes eleven problems
For most of them I have a working program (a real understandable code snippet with less than 10 instructions that solves all the cases), others have been discussed on Kaggle. Some may be open to interpretation, but bear in mind that when you have a working formal description of the solution (a piece of code) this should replace human preconceptions. Take 7039b2d7.json for example (case 3). If you think it is okay the way it is, you should ask yourself why the two inexistent rectangles to the left of the picture are blue? Our bias makes them the same color as the others, but. If they don't exist, they have no color. Take the simpler solution, just remove them, there is nothing to the left of the leftmost yellow line.